reorganise auth folder a bit
