--- src/auth/errors.js
@@ -1,5 +1,0 @@
-const unauthorized = "The email or password you have entered is invalid.";
-
-module.exports = {
-  unauthorized
-};
