+++ src/auth/routes/logout.js
@@ -1,4 +1,4 @@
-const { removeToken } = require("../../middleware/whiteList");
+const { removeToken } = require("../../../middleware/whiteList");
 const boom = require("boom");
 
 const logout = (req, res, next) => {
