+++ src/auth/routes/register.js
