+++ src/auth/tests/login.test.js
@@ -2,21 +2,20 @@
  * @jest-environment node
  */
 
-// Testing Registration
+// Testing Login
 const axios = require("axios");
-const { port } = require("../../config");
+const { port } = require("../../../config");
 const uuidv4 = require("uuid/v4");
 
-describe("Registration", () => {
+describe("Login", () => {
   it("should reject an empty submission with status 422", async () => {
     const expectedResponseCode = 422;
     const expectedResponseText = "Unprocessable Entity";
-
     try {
       const response = await axios({
-        url: `http://localhost:${port}/auth/register`,
+        url: `http://localhost:${port}/auth/login`,
         method: "POST",
-        data: null
+        data: { email: `${uuidv4()}@testing.com` }
       });
       if (response) throw new Error();
     } catch (error) {
@@ -28,12 +27,11 @@ describe("Registration", () => {
   it("should reject an empty password submission with status 422", async () => {
     const expectedResponseCode = 422;
     const expectedResponseText = "Unprocessable Entity";
-
     try {
       const response = await axios({
-        url: `http://localhost:${port}/auth/register`,
+        url: `http://localhost:${port}/auth/login`,
         method: "POST",
-        data: { email: `${uuidv4()}@testing.com`, confirmPassword: uuidv4() }
+        data: { email: `${uuidv4()}@testing.com` }
       });
       if (response) throw new Error();
     } catch (error) {
@@ -45,31 +43,11 @@ describe("Registration", () => {
   it("should reject an empty email submission with status 422", async () => {
     const expectedResponseCode = 422;
     const expectedResponseText = "Unprocessable Entity";
-
-    const password = uuidv4();
-
     try {
       const response = await axios({
-        url: `http://localhost:${port}/auth/register`,
+        url: `http://localhost:${port}/auth/login`,
         method: "POST",
-        data: { password, confirmPassword: password }
-      });
-      if (response) throw new Error();
-    } catch (error) {
-      expect(error.response.status).toEqual(expectedResponseCode);
-      expect(error.response.statusText).toEqual(expectedResponseText);
-    }
-  });
-
-  it("should reject an empty confirmPassword submission with status 422", async () => {
-    const expectedResponseCode = 422;
-    const expectedResponseText = "Unprocessable Entity";
-
-    try {
-      const response = await axios({
-        url: `http://localhost:${port}/auth/register`,
-        method: "POST",
-        data: { email: `${uuidv4()}@test.com`, password: uuidv4() }
+        data: { password: uuidv4() }
       });
       if (response) throw new Error();
     } catch (error) {
@@ -81,17 +59,12 @@ describe("Registration", () => {
   it("should reject an email of an incorrect type with status 422", async () => {
     const expectedResponseCode = 422;
     const expectedResponseText = "Unprocessable Entity";
-    const password = uuidv4();
 
     try {
       const response = await axios({
-        url: `http://localhost:${port}/auth/register`,
+        url: `http://localhost:${port}/auth/login`,
         method: "POST",
-        data: {
-          email: { object: uuidv4() },
-          password,
-          confirmPassword: password
-        }
+        data: { email: { object: uuidv4() }, password: uuidv4() }
       });
       if (response) throw new Error();
     } catch (error) {
@@ -103,13 +76,12 @@ describe("Registration", () => {
   it("should reject a password of an incorrect type with status 422", async () => {
     const expectedResponseCode = 422;
     const expectedResponseText = "Unprocessable Entity";
-    const password = [uuidv4()];
 
     try {
       const response = await axios({
-        url: `http://localhost:${port}/auth/register`,
+        url: `http://localhost:${port}/auth/login`,
         method: "POST",
-        data: { email: "543", password, confirmPassword: password }
+        data: { email: "543", password: [uuidv4()] }
       });
       if (response) throw new Error();
     } catch (error) {
@@ -118,35 +90,15 @@ describe("Registration", () => {
     }
   });
 
-  it("should reject an invalid email address with status 422", async () => {
-    const expectedResponseCode = 422;
-    const expectedResponseText = "Unprocessable Entity";
-    const password = uuidv4();
+  it("should reject an incorrect email or password with status 401", async () => {
+    const expectedResponseCode = 401;
+    const expectedResponseText = "Unauthorized";
 
     try {
       const response = await axios({
-        url: `http://localhost:${port}/auth/register`,
+        url: `http://localhost:${port}/auth/login`,
         method: "POST",
-        data: { email: uuidv4(), password, confirmPassword: password }
-      });
-      if (response) throw new Error();
-    } catch (error) {
-      expect(error.response.status).toEqual(expectedResponseCode);
-      expect(error.response.statusText).toEqual(expectedResponseText);
-    }
-  });
-  it("should reject if password and confirmPassword are not equal", async () => {
-    const expectedResponseCode = 422;
-    const expectedResponseText = "Unprocessable Entity";
-    const email = `${uuidv4()}@testing.com`;
-    const password = uuidv4();
-    const confirmPassword = uuidv4();
-
-    try {
-      const response = await axios({
-        url: `http://localhost:${port}/auth/register`,
-        method: "POST",
-        data: { email, password, confirmPassword }
+        data: { email: uuidv4(), password: uuidv4() }
       });
       if (response) throw new Error();
     } catch (error) {
@@ -155,45 +107,58 @@ describe("Registration", () => {
     }
   });
 
-  it("should accept a valid email address with status 201", async () => {
-    const expectedResponseCode = 201;
-    const password = uuidv4();
-
-    const response = await axios({
-      url: `http://localhost:${port}/auth/register`,
-      method: "POST",
-      data: {
-        email: `${uuidv4()}@testing.com`,
-        password,
-        confirmPassword: password
-      }
-    });
-    expect(response.status).toEqual(expectedResponseCode);
-  });
-  it("should reject if the email address has already been taken", () => {
-    const expectedResponseCode = 422;
-    const expectedResponseText = "Unprocessable Entity";
+  it("should reject a correct email with an incorrect password with status 401", async () => {
+    const expectedResponseCode = 401;
+    const expectedResponseText = "Unauthorized";
 
     const email = `${uuidv4()}@testing.com`;
     const password = uuidv4();
 
-    axios({
+    // create account.
+    const regResponse = await axios({
       url: `http://localhost:${port}/auth/register`,
       method: "POST",
       data: { email, password, confirmPassword: password }
-    }).then(() => {
-      axios({
-        url: `http://localhost:${port}/auth/register`,
-        method: "POST",
-        data: { email, password, confirmPassword: password }
-      })
-        .then(response => {
-          if (response) throw new Error();
-        })
-        .catch(error => {
-          expect(error.response.status).toEqual(expectedResponseCode);
-          expect(error.response.statusText).toEqual(expectedResponseText);
-        });
     });
+
+    // Try to log in with random password.
+    try {
+      const loginResponse = await axios({
+        url: `http://localhost:${port}/auth/login`,
+        method: "POST",
+        data: { email, password: uuidv4() }
+      });
+      if (loginResponse) throw new Error();
+    } catch (error) {
+      expect(error.response.status).toEqual(expectedResponseCode);
+      expect(error.response.statusText).toEqual(expectedResponseText);
+    }
+  });
+
+  it("should accept a correct email and password, responding with a 201", async () => {
+    const expectedResponseCode = 201;
+    const expectedResponseProperties = ["success", "token"];
+
+    const email = `${uuidv4()}@testing.com`;
+    const password = uuidv4();
+
+    // create account.
+    const regResponse = await axios({
+      url: `http://localhost:${port}/auth/register`,
+      method: "POST",
+      data: { email, password, confirmPassword: password }
+    });
+
+    // Try to log in with correct password
+    const loginResponse = await axios({
+      url: `http://localhost:${port}/auth/login`,
+      method: "POST",
+      data: { email, password }
+    });
+    const responseProperties = Object.keys(loginResponse.data);
+    expect(loginResponse.status).toEqual(expectedResponseCode);
+    expect(responseProperties).toEqual(
+      expect.arrayContaining(expectedResponseProperties)
+    );
   });
 });
